requirements.txt
asgiref==3.5.2
boto3==1.24.52
botocore==1.27.52
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
gunicorn==20.0.4
jmespath==1.0.1
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
